Move language resource reading in another file.
