+++ program/noMeiryoUI.vcxproj
@@ -188,6 +188,7 @@
     <ClCompile Include="country\korea.cpp" />
     <ClCompile Include="FontSel.cpp" />
     <ClCompile Include="iniReader.cpp" />
+    <ClCompile Include="langresource.cpp" />
     <ClCompile Include="noMeiryoUI.cpp" />
     <ClCompile Include="stdafx.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug_2008|Win32'">Create</PrecompiledHeader>
@@ -212,6 +213,7 @@
     <ClInclude Include="country\korea.h" />
     <ClInclude Include="FontSel.h" />
     <ClInclude Include="iniReader.h" />
+    <ClInclude Include="langresource.h" />
     <ClInclude Include="noMeiryoUI.h" />
     <ClInclude Include="Resource.h" />
     <ClInclude Include="stdafx.h" />
