+++ program/noMeiryoUI.vcxproj.filters
@@ -78,6 +78,9 @@
     <ClCompile Include="country\default.cpp">
       <Filter>ソース ファイル\country</Filter>
     </ClCompile>
+    <ClCompile Include="langresource.cpp">
+      <Filter>ソース ファイル</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="FontSel.h">
@@ -145,6 +148,9 @@
     </ClInclude>
     <ClInclude Include="country\default.h">
       <Filter>ヘッダー ファイル\country</Filter>
+    </ClInclude>
+    <ClInclude Include="langresource.h">
+      <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
